Refactor grid coordinate utilities into separate GridUtils module
